debt - move perfview editor to browser
