+++ src/vs/workbench/contrib/performance/browser/perfviewEditor.ts
@@ -421,6 +421,5 @@ class MarkdownBuilder {
 			});
 			this.value += '|\n';
 		});
-
 	}
 }
